+++ R/families.R
@@ -13,7 +13,7 @@
 #'   variable be used in the model. Currently, the following families are
 #'   supported: \code{gaussian}, \code{student}, \code{binomial},
 #'   \code{bernoulli}, \code{poisson}, \code{negbinomial}, \code{geometric},
-#'   \code{Gamma}, \code{skew_normal}, \code{lognormal},
+#'   \code{discrete_weibull}, \code{Gamma}, \code{skew_normal}, \code{lognormal},
 #'   \code{shifted_lognormal}, \code{exgaussian}, \code{wiener},
 #'   \code{inverse.gaussian}, \code{exponential}, \code{weibull},
 #'   \code{frechet}, \code{Beta}, \code{dirichlet}, \code{von_mises},
@@ -437,6 +437,14 @@ negbinomial <- function(link = "log", link_shape = "log") {
 geometric <- function(link = "log") {
   slink <- substitute(link)
   .brmsfamily("geometric", link = link, slink = slink)
+}
+
+#' @rdname brmsfamily
+#' @export
+discrete_weibull <- function(link = "logit", link_shape = "log") {
+  slink <- substitute(link)
+  .brmsfamily("discrete_weibull", link = link, slink = slink,
+              link_shape = link_shape)
 }
 
 #' @rdname brmsfamily