+++ tests/testthat/tests.predict.R
@@ -147,6 +147,9 @@ test_that("predict for count and survival models runs without errors", {
   pred <- brms:::predict_binomial(i, draws = draws)
   expect_equal(length(pred), ns)
   
+  pred <- brms:::predict_discrete_weibull(i, draws = draws)
+  expect_equal(length(pred), ns)
+  
   draws$dpars$mu <- exp(draws$dpars$eta)
   pred <- brms:::predict_poisson(i, draws = draws)
   expect_equal(length(pred), ns)